bugfix(avl): wrong proof after updating a key
When updating one of the keys the proof was no longer valid. It turns
our that the node's value wasn't updated, but the test suite did not
catch that bug... This commit fixes the issue by correctly updating the
value and triggering an update of the local hash on value update.